put dh_set_pg_dhparam() in own c file
